Remove user model
