+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_02_05_110819) do
+ActiveRecord::Schema.define(version: 2021_02_05_101032) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
@@ -34,23 +34,18 @@ ActiveRecord::Schema.define(version: 2021_02_05_110819) do
 
   create_table "accounts", force: :cascade do |t|
     t.string "email", null: false
-    t.index ["email"], name: "index_accounts_on_email", unique: true
-  end
-
-  create_table "users", force: :cascade do |t|
-    t.string "login"
+    t.string "username", null: false
     t.text "description"
     t.string "profile_photo"
     t.integer "followers"
     t.integer "following"
-    t.bigint "account_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.index ["account_id"], name: "index_users_on_account_id"
+    t.index ["email"], name: "index_accounts_on_email", unique: true
+    t.index ["username"], name: "index_accounts_on_username", unique: true
   end
 
   add_foreign_key "account_login_change_keys", "accounts", column: "id"
   add_foreign_key "account_password_hashes", "accounts", column: "id"
   add_foreign_key "account_password_reset_keys", "accounts", column: "id"
-  add_foreign_key "users", "accounts"
 end
